Recent Message Cells has been configured.
